feat: create User, record part
