--- src/main/java/Dongguk/Blinking_AI/BlinkingAiApplication.java
@@ -1,13 +1,0 @@
-package Dongguk.Blinking_AI;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BlinkingAiApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BlinkingAiApplication.class, args);
-	}
-
-}
