--- src/main/java/Dongguk/Blinking_AI/user/dto/response/MyPageUpdateResponse.java
@@ -1,7 +1,0 @@
-package Dongguk.Blinking_AI.user.dto.response;
-
-public record MyPageUpdateResponse (
-        Integer frequencyAlert,
-        Integer displayType,
-        Integer alertType
-){}
