--- src/main/java/Dongguk/Blinking_AI/user/repository/UserRepository.java
@@ -1,8 +1,0 @@
-package Dongguk.Blinking_AI.user.repository;
-
-import Dongguk.Blinking_AI.user.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-}
