+++ src/main/java/Dongguk/Blinking_Eyes/BlinkingAiApplication.java
@@ -1,0 +1,13 @@
+package Dongguk.Blinking_Eyes;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BlinkingAiApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BlinkingAiApplication.class, args);
+	}
+
+}
