+++ src/main/java/Dongguk/Blinking_Eyes/blink/dto/response/MonthlyStaticsResponse.java
@@ -1,0 +1,13 @@
+package Dongguk.Blinking_Eyes.blink.dto.response;
+
+import Dongguk.Blinking_Eyes.blink.domain.DayCount;
+import lombok.Builder;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Builder
+public record MonthlyStaticsResponse (
+        String month,
+        List<DayCount> list
+){}
