+++ src/main/java/Dongguk/Blinking_Eyes/blink/repository/HealthRecordRepository.java
@@ -1,0 +1,7 @@
+package Dongguk.Blinking_Eyes.blink.repository;
+
+import Dongguk.Blinking_Eyes.blink.domain.HealthRecord;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface HealthRecordRepository extends JpaRepository<HealthRecord, Long> {
+}
