+++ src/main/java/Dongguk/Blinking_Eyes/exception/CommonException.java
@@ -1,0 +1,13 @@
+package Dongguk.Blinking_Eyes.exception;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class CommonException extends RuntimeException {
+    private final ErrorCode errorCode;
+    public String getMessage() {
+        return this.errorCode.getMessage();
+    }
+}
