+++ src/main/java/Dongguk/Blinking_Eyes/user/dto/request/MyPageUpdateRequest.java
@@ -1,0 +1,7 @@
+package Dongguk.Blinking_Eyes.user.dto.request;
+
+public record MyPageUpdateRequest (
+    Integer eyeSensitivity,
+    Integer alarmPeriod,
+    Boolean isSound
+){}
