+++ src/main/resources/application.yml
@@ -1,8 +1,8 @@
 spring:
   datasource:
-    url: "jdbc:mysql://localhost/dtc"
+    url: "jdbc:mysql://localhost:3306/blink"
     username: "root"
-    password: ""
+    password: "0302"
     driver-class-name: com.mysql.cj.jdbc.Driver
   jpa:
     hibernate:
@@ -12,3 +12,5 @@ spring:
         show_sql: true
         format_sql: true
         dialect: org.hibernate.dialect.MySQL8Dialect
+
+
