--- src/test/java/Dongguk/Blinking_AI/BlinkingAiApplicationTests.java
@@ -1,13 +1,0 @@
-package Dongguk.Blinking_AI;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BlinkingAiApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
