+++ src/test/java/Dongguk/Blinking_Eyes/BlinkingAiApplicationTests.java
@@ -1,0 +1,13 @@
+package Dongguk.Blinking_Eyes;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BlinkingAiApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
